refactor: migrate 4 backlog modules from scripts/ to src/backlog/ (#344)
T9-T12 of ADR-0010 implementation, landed as one slice because
scripts/continuous_burndown.py imports all four under a single
try/except block — splitting these into four commits leaves the
codebase un-importable between commits.

Migrated:
- scripts/backlog_groomer.py -> src/backlog/backlog_groomer.py
- scripts/ci_health_monitor.py -> src/backlog/ci_health_monitor.py
- scripts/migrate_backlog_to_github.py -> src/backlog/migrate_backlog_to_github.py
- scripts/validate_documentation_accuracy.py -> src/backlog/validate_documentation_accuracy.py

Caller updated:
- scripts/continuous_burndown.py — primary import branch now uses
  src.backlog.*; the bare-name fallback retained for direct script
  execution.

Fixes repo-root resolution in DocumentationValidator.__init__:
src/backlog/validate_documentation_accuracy.py uses parent.parent.parent.

Refs #344
